Remove excess chaining
When I chain each command in the test (cypress@4.3.0 and @testing-library/cypress@5.3.1), I get the error message in the browser console:
Error: @testing-library/cypress will eventually only use previous subjects when queries are added to a chain of commands. We've detected an instance where the this functionality failed, but the old functionality passed (so your test may break in a future version). Please use cy.findByText(/^1$/) instead of continuing from a previous chain.